Tidy up marshallers/unmarshallers
