+++ consensus/pbft/message.go
@@ -1,7 +1,6 @@
 package pbft
 
 import (
-	"encoding/json"
 	"fmt"
 	"time"
 
@@ -87,72 +86,4 @@ type NewView struct {
 	View        uint                   `json:"view"`
 	Messages    map[peer.ID]ViewChange `json:"messages"`
 	PrePrepares []PrePrepare           `json:"preprepares"`
-}
-
-// messageRecord is used as an interim format to supplement the original type with its type.
-// Useful for serialization to automatically include the message type field.
-type messageRecord struct {
-	Type MessageType     `json:"type"`
-	Data json.RawMessage `json:"data"`
-}
-
-func unpackMessage(payload []byte) (any, error) {
-
-	var msg messageRecord
-	err := json.Unmarshal(payload, &msg)
-	if err != nil {
-		return nil, fmt.Errorf("could not unpack base message: %w", err)
-	}
-
-	switch msg.Type {
-	case MessageRequest:
-		var request Request
-		err = json.Unmarshal(msg.Data, &request)
-		if err != nil {
-			return nil, fmt.Errorf("could not unpack request: %w", err)
-		}
-		return request, nil
-
-	case MessagePrePrepare:
-		var preprepare PrePrepare
-		err = json.Unmarshal(msg.Data, &preprepare)
-		if err != nil {
-			return nil, fmt.Errorf("could not unpack pre-prepare message: %w", err)
-		}
-		return preprepare, nil
-
-	case MessagePrepare:
-		var prepare Prepare
-		err = json.Unmarshal(msg.Data, &prepare)
-		if err != nil {
-			return nil, fmt.Errorf("could not unpack prepare message: %w", err)
-		}
-		return prepare, nil
-
-	case MessageCommit:
-		var commit Commit
-		err = json.Unmarshal(msg.Data, &commit)
-		if err != nil {
-			return nil, fmt.Errorf("could not unpack commit message: %w", err)
-		}
-		return commit, nil
-
-	case MessageViewChange:
-		var viewChange ViewChange
-		err = json.Unmarshal(msg.Data, &viewChange)
-		if err != nil {
-			return nil, fmt.Errorf("could not unpack view change message: %w", err)
-		}
-		return viewChange, nil
-
-	case MessageNewView:
-		var newView NewView
-		err = json.Unmarshal(msg.Data, &newView)
-		if err != nil {
-			return nil, fmt.Errorf("could not unpack new view message: %w", err)
-		}
-		return newView, nil
-	}
-
-	return nil, fmt.Errorf("unexpected message type (type: %v)", msg.Type)
 }
